requirements.txt
selenium==4.25.0
pandas==2.0.3
python-dotenv==1.0.1
requests==2.22.0
tqdm==4.66.5
